+++ client/components/SingleEpisode/SingleEpisode.js
@@ -19,7 +19,6 @@ class SingleEpisode extends React.Component {
   async componentDidMount() {
     const episodeId = this.props.match.params.episodeId
     try {
-      console.log('Test')
       await this.props.getCurrentEpisode(episodeId)
     } catch (err) {
       console.log('Error!', err)
@@ -52,7 +51,9 @@ class SingleEpisode extends React.Component {
           <Route
             path={`/episodes/${episodeId}/cast`}
             render={() => (
-              <div className="singleEpisodeDeets mt-4 elevatedCard red-trim flexDown">
+              <div className="singleEpisodeDeets-base mt-4 elevatedCard flexDown">
+              <div className="singleEpisodeDeets-contents mt-4flexDown">
+              <h2 className="text-center">Cast:</h2>
                 {characters.map(character => (
                   <h4 key={character.id}>
                     {' '}
@@ -60,10 +61,11 @@ class SingleEpisode extends React.Component {
                   </h4>
                 ))}
               </div>
+              </div>
             )}
           />
+          </div>
         </div>
-      </div>
     )
   }
 }
